Completed Goodreads scraper. Gets review value and puts in UI. Also organized files a bit
